naprawa sponsora

+++ Donator.java
@@ -5,7 +5,7 @@ public class Donator{
 	private final String lastName;
 	private double allMoneyDonated=0.0;
 	
-	public Donator(String name, String lastName, int initialDonation) {
+	public Donator(String name, String lastName, double initialDonation) {
 		this.name = name;
 		this.lastName = lastName;
 		allMoneyDonated += initialDonation;
